pyproject.toml
[project]
name = "v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "customtkinter>=5.2.2",
    "python-dateutil>=2.9",
    "tkcalendar>=1.6.1",
]
